pyproject.toml
[project]
name = "fashion-hub-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Vaghasiya, Ravibhai (EXTERN: IAV GmbH)",email = "ravibhai.vaghasiya@iav.de"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ruff (>=0.11.11,<0.12.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "fastapi",
    "uvicorn",
    "pydantic-settings",
    "sqlalchemy",
    "python-dotenv",
    "psycopg2",
    "pandas",
    "python-multipart",
    "strenum",
    "jinja2",
    "langchain",
    "text-generation",
    "langchain-experimental",
    "pgvector",
    "pypdf",
    "ahocorasick-rs",
    "sqlalchemy-utils",
    "langchain-openai",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (<=4.0.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "fashion_hub_backend", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.110.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
pandas = "^2.1.3"
python-multipart = "^0.0.6"
strenum = "^0.4.15"
jinja2 = "^3.1.2"
langchain = "^0.2.16"
text-generation = ">=0.7.0"
langchain-experimental = "^0.0.65"
pgvector = "^0.2.5"
pypdf = "^4.3.1"
ahocorasick-rs = "^0.22.0"
sqlalchemy-utils = "^0.41.2"
langchain-openai = "^0.1.23"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "F",    # pycodestyle
    "E",    # Pyflakes
    "W",    # Warnings
    "U",    # pyupgrade
    "I",    # isort
    "N",    # pep8-naming
    "PL",   # Pylint
    "B",    # flake8-bugbear
    "RUF",  # Ruff-specific codes
    "PERF", # Performance-related issues
    "SIM",  # flake8-simplify
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "C901", # maccabe
    "S",    # flake8-bandit
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "INT",  # flake8-gettext
    "FLY002", # flynt
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
]
ignore = [
    "PLR2004",  # Magic value used in comparison, consider replacing `X` with a constant variable
    "ISC001",  # to let formatter run
    "N806",
    "B008",
    "PLR0913",
    "RUF012",
    "TC001",
    "F821"
]
exclude = [
    ".git",
    ".venv",
    ".pyenv",
    "venv",
    "__pycache__",
    "__target__",
    "dist",
]